Move OldMujocoResettableWrapper to lookback (only place it is used)
