+++ src/aprl/common/mujoco.py
@@ -4,8 +4,6 @@ from collections import namedtuple
 import gym
 import numpy as np
 
-from aprl.envs.multi_agent import MultiWrapper
-
 
 # TODO: Cythonize
 class MujocoState(namedtuple('MujocoStateBase', 'qpos qvel')):
@@ -46,76 +44,3 @@ class ResettableEnv(gym.Env):
         """Restores the environment to a previously saved state.
         :param x: return value of a previous call to get_state()."""
         pass
-
-
-class OldMujocoResettableWrapper(ResettableEnv, MultiWrapper):
-    """Converts a MujocoEnv into a ResettableEnv.
-
-    Specifically designed to handle getting and setting states with Mujoco 1.31 / mujoco_py 0.5.7
-    Note all MuJoCo environments are resettable."""
-    def __init__(self, env):
-        """Wraps a MujocoEnv, adding get_state and set_state methods.
-        :param env: a MujocoEnv. NOTE: it must not be wrapped in a TimeLimit."""
-        if hasattr(env, '_max_episode_steps'):
-            raise TypeError('Environment must not have a time limit '
-                            '(try passing in env.unwrapped instead).')
-        gym.Wrapper.__init__(self, env)
-        self.sim = env.unwrapped.env_scene
-
-    def get_sim_data(self):
-        """Get the contents of an MjData object in dict form for the environment
-
-        :return: (dict<str, [float]) state_dict
-        """
-        state_dict = {}
-        for k, v in type(self.sim.data._wrapped.contents).__dict__['_fields_']:
-            if k not in ['contact', 'buffer']:
-                state_dict[k] = getattr(self.sim.data, k)
-        return state_dict
-
-    def get_radius(self):
-        """Get the radius of the environment, assuming it has one.
-
-        :return: (float) radius
-        """
-        return self.env.env.RADIUS
-
-    def get_state(self):
-        """Serializes the qpos and qvel state of the MuJoCo emulator.
-
-        :return: ([float]) state
-        """
-        state = MujocoState.from_mjdata(self.sim.data).flatten()
-        return state
-
-    def set_state(self, x, forward=True):
-        """Restores qpos and qvel, calling forward() to derive other values.
-
-        :param forward (bool) whether to call forward on the environment.
-        """
-        state = MujocoState.from_flattened(x, self.sim)
-        state.set_mjdata(self.sim.data)
-        if forward:
-            self.sim.model.forward()  # put mjData in consistent state
-
-    def set_radius(self, radius):
-        """Set the radius of this environment.
-
-        :param radius (float) size of environment, if it exists"""
-        self.env.env.RADIUS = radius
-        self.env.env._set_geom_radius()
-
-    def set_sim_data(self, sim_data):
-        """Set the fields of the MjData object of this environment
-
-        :param sim_data (dict<str, [float]) dict with MjData fields extracted using get_state"""
-        for k, v in sim_data.items():
-            setattr(self.sim.data, k, v)
-
-    def reset(self):
-        """See base class."""
-        return self.env.reset()
-
-    def step(self, a):
-        """See base class."""
-        return self.env.step(a)
